storage_controller: enforce generations in import upcalls
Problem

Import up-calls did not enforce the usage of the latest generation.
The import might have finished in one previous generation, but not
in the latest one. Hence, the controller might try to activate a
timeline before it is ready. In theory, that would be fine, but
it's tricky to reason about.

Summary of Changes

Pageserver provides the current generation in the upcall to the storage
controller and the later validates the generation. If the generation is
stale, we return an error which stops progress of the import job. Note
that the import job will retry the upcall until the stale location is
detached.

Closes https://github.com/neondatabase/neon/issues/11884